Optimize ut
